# pos-backend/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: POS Papelería — API REST
  version: 1.0.0
  description: |
    API del sistema de punto de venta para papelería.
    Desarrollada con Node.js/Express + Supabase (PostgreSQL).

    ## Autenticación
    Todos los endpoints (salvo `/auth/login`) requieren un **JWT Bearer token**
    obtenido desde `POST /api/auth/login`.

    ```
    Authorization: Bearer <token>
    ```

    ## Roles
    | Rol | Permisos |
    |---|---|
    | `cajero` | ventas, consulta de productos e inventario |
    | `administrador` | todo lo anterior + productos, inventario, corte de caja |

    ## Formato de respuesta exitosa
    ```json
    { "success": true, "data": { ... } }
    ```

    ## Formato de error
    ```json
    { "success": false, "error": "Mensaje descriptivo" }
    ```
servers:
  - url: http://localhost:3000/api
    description: Desarrollo local
  - url: http://192.168.1.100:3000/api
    description: Red local ITH

# ──────────────────────────────────────────────────────────────────────────────
# Security scheme
# ──────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido desde POST /auth/login. Expira en 10 horas.

  # ── Schemas reutilizables ────────────────────────────────────────────────────
  schemas:
...
    Error400:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Mensaje de validación" }
    Error401:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Token inválido o expirado" }
    Error403:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Se requiere rol administrador" }
    Error404:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:   { type: string,  example: "Recurso no encontrado" }
...
    Producto:
      type: object
      properties:
        id_producto:   { type: integer, example: 7 }
        codigo_barras: { type: string,  example: "7501000000007" }
        nombre:        { type: string,  example: "Cuaderno universitario 100 hojas" }
        descripcion:   { type: string,  example: "Raya, pasta dura", nullable: true }
        id_categoria:  { type: integer, example: 1, nullable: true }
        precio_compra: { type: number,  format: decimal, example: 18.50 }
        precio_venta:  { type: number,  format: decimal, example: 35.00 }
        stock_actual:  { type: integer, example: 48 }
        stock_minimo:  { type: integer, example: 10 }
        activo:        { type: boolean, example: true }
        creado_en:     { type: string,  format: date-time }
        actualizado_en: { type: string,  format: date-time }
...
security:
  - bearerAuth: []

# ──────────────────────────────────────────────────────────────────────────────
# Tags
# ──────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /productos:
    post:
      tags: [Productos]
      summary: Crear producto
      description: >
        Registra un nuevo producto en el catálogo.
        Requiere autenticación (cualquier rol).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, codigo_barras, precio_compra, precio_venta]
              properties:
                nombre:        { type: string,  example: "Cuaderno universitario 100 hojas" }
                codigo_barras: { type: string,  example: "7501000000007" }
                descripcion:   { type: string,  example: "Raya, pasta dura" }
                id_categoria:  { type: integer, example: 1 }
                precio_compra: { type: number,  example: 18.50 }
                precio_venta:  { type: number,  example: 35.00 }
                stock_actual:  { type: integer, example: 50,  default: 0 }
                stock_minimo:  { type: integer, example: 10,  default: 5 }
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: '#/components/schemas/Producto' }
        '400':
          description: Validación fallida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error400' }
              examples:
                codigo_duplicado:
                  value: { success: false, error: "El código de barras ya existe" }
                precio_invalido:
                  value: { success: false, error: "El precio de venta debe ser mayor a 0" }
        '401': { description: No autenticado,    content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }

    get:
      tags: [Productos]
      summary: Listar / buscar productos
      description: >
        Devuelve todos los productos activos.
        El parámetro `q` filtra por nombre o código de barras (búsqueda parcial, case-insensitive).
      parameters:
        - name: q
          in: query
          required: false
          description: Texto para buscar en nombre o código de barras
          schema:
            type: string
            example: "cuaderno"
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Producto' }
              example:
                success: true
                data:
                  - id_producto: 7
                    codigo_barras: "7501000000007"
                    nombre: "Cuaderno universitario 100 hojas"
                    precio_venta: 35.00
                    stock_actual: 48
                    stock_minimo: 10
                    activo: true
        '401': { description: No autenticado, content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
  /productos/barcode/{codigo}:
    get:
      tags: [Productos]
      summary: Buscar por código de barras
      description: >
        Localiza un producto activo por su código de barras exacto.
        Utilizado por el escáner del POS para agregar productos al carrito.
      parameters:
        - name: codigo
          in: path
          required: true
          schema: { type: string, example: "7501000000007" }
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: '#/components/schemas/Producto' }
        '404':
          description: Código de barras no registrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error404' }
              example: { success: false, error: "Producto no encontrado" }
        '401': { description: No autenticado, content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
  /productos/{id}:
    get:
      tags: [Productos]
      summary: Obtener producto por ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 7 }
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: '#/components/schemas/Producto' }
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error404' }
        '401': { description: No autenticado, content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }

    put:
      tags: [Productos]
      summary: Actualizar producto
      description: >
        Actualización parcial (PATCH semántico implementado con PUT).
        Solo enviar los campos que se desean modificar.
        **Requiere rol administrador.**
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 7 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:        { type: string,  example: "Cuaderno universitario 200 hojas" }
                precio_venta:  { type: number,  example: 45.00 }
                stock_minimo:  { type: integer, example: 15 }
                activo:        { type: boolean, example: true }
            examples:
              actualizar_precio:
                summary: Cambiar precio de venta
                value: { precio_venta: 45.00 }
              dar_de_baja:
                summary: Desactivar producto
                value: { activo: false }
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: '#/components/schemas/Producto' }
        '400': { description: Validación fallida,  content: { application/json: { schema: { $ref: '#/components/schemas/Error400' } } } }
        '401': { description: No autenticado,       content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } } }
        '403': { description: Rol insuficiente,     content: { application/json: { schema: { $ref: '#/components/schemas/Error403' } } } }
        '404': { description: Producto no encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/Error404' } } } }

  # ═══════════════════════════════════════════════════════════════════════════
  # VENTAS
  # ═══════════════════════════════════════════════════════════════════════════
...